Edited errors 1
