+++ js/create-big-picture.js
@@ -1,4 +1,4 @@
-import { createComments } from './createComments.js';
+import { createComments } from './create-comments.js';
 
 const bigPicture = document.querySelector('.big-picture');
 const bigPictureLikes = bigPicture.querySelector('.likes-count');
@@ -7,7 +7,7 @@ const bigPictureDesctiption = bigPicture.querySelector('.social__caption');
 const socialComments = document.querySelector('.social__comments');
 const socialComment = document.querySelector('.social__comment');
 
-const createBigPicture = function (post) {
+const createBigPicture = (post) => {
   bigPicture.querySelector('.big-picture__img img').src = post.url;
   bigPicture.querySelector('.big-picture__img img').alt = post.description;
   bigPictureLikes.textContent = post.likes;
