--- js/getCurrentPostId.js
@@ -1,9 +1,0 @@
-import { getNumber } from './util.js';
-
-function getCurrentPostId(currentPicture) {
-  const currentPictureSrc = currentPicture.getAttribute('src');
-  const currentPictureId = getNumber(currentPictureSrc);
-  return currentPictureId;
-}
-
-export { getCurrentPostId };
