--- js/getObjectData.js
@@ -1,6 +1,0 @@
-function getObjectData(pictureId, data) {
-  const currentPostData = data.find((current) => current.id === pictureId);
-  return currentPostData;
-}
-
-export { getObjectData };
