+++ js/main.js
@@ -1,11 +1,9 @@
-import './thumbnails.js';
-import './openBigPicture.js';
 import { createThumbnails } from './thumbnails.js';
 import { createPosts } from './data.js';
-import { addListener } from './openBigPicture.js';
+import { addPicturesClickListener } from './open-big-picture.js';
 
 const dataPosts = createPosts();
 createThumbnails(dataPosts);
-addListener();
+addPicturesClickListener();
 
 export {dataPosts};
