+++ js/util.js
@@ -9,7 +9,7 @@ const getRandomArrayElement = (elements) => elements[getRandomInteger(0, element
 
 export {getRandomInteger, getRandomArrayElement};
 
-const getNumber = function (string) {
+const getNumber = (string) => {
   if (string === undefined || string === null) {
     return NaN;
   }
@@ -22,3 +22,18 @@ const getNumber = function (string) {
 };
 
 export { getNumber };
+
+const getObjectData = (pictureId, data) => {
+  const currentPostData = data.find((current) => current.id === pictureId);
+  return currentPostData;
+};
+
+export { getObjectData };
+
+const getCurrentPostId = (currentPicture) => {
+  const currentPictureSrc = currentPicture.getAttribute('src');
+  const currentPictureId = getNumber(currentPictureSrc);
+  return currentPictureId;
+};
+
+export { getCurrentPostId };
